+++ src/prelude.rs
@@ -11,6 +11,7 @@
 //! convenience.
 
 use std::iter;
+use std::panic;
 
 use futures::{Future, Sink, Stream};
 
@@ -66,7 +67,7 @@ pub trait CoroutineFuture: Sized {
     /// # }
     /// ```
     fn coro_wait(self) -> Result<Self::Item, Self::Error> {
-        self.coro_wait_cleanup().unwrap()
+       self.coro_wait_cleanup().unwrap_or_else(|_| panic::resume_unwind(Box::new(Dropped)))
     }
 
     /// A coroutine aware wait on the result that doesn't panic.